LibWeb: Delete the C++ query parser
Remove the legacy media, supports, container, style, and sizes query
grammar after all production callers moved to Rust. Retain only the C++
evaluation facades and conversion helpers used by the Rust query arena.
